More refactoring
Split options window into new file, fixed bug in js where both buttons advanced the instruction. Also added counter to instruction.
