+++ src/options_window.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void init_options_window();
