+++ src/panoramio.js
@@ -30,11 +30,11 @@ function sendCurrInstruction(destination) {
         if (destination) {
             Pebble.sendAppMessage({
                 'DESTINATION': String(destination),
-                'DIRECTIONS': stripHtmlTags(String(navInstructions[instructionCounter].html_instructions))
+                'DIRECTIONS': stripHtmlTags(String(instructionCounter+1) + ": " + String(navInstructions[instructionCounter].html_instructions))
             });
         } else {
             Pebble.sendAppMessage({
-                'DIRECTIONS': stripHtmlTags(String(navInstructions[instructionCounter].html_instructions))
+                'DIRECTIONS': stripHtmlTags(String(instructionCounter+1) + ": " + String(navInstructions[instructionCounter].html_instructions))
             });
         }
         console.log("sent instructions: " + stripHtmlTags(navInstructions[instructionCounter].html_instructions));
@@ -120,10 +120,14 @@ Pebble.addEventListener("appmessage", function(e) {
     if ('INIT_DIRECTIONS' in e.payload) {
         navigator.geolocation.getCurrentPosition(locationSuccess, locationError, locationOptions);
     } else if ('NEXT_INSTRUCTION' in e.payload) {
-        instructionCounter++;
-        sendCurrInstruction();
+        if (instructionCounter < navInstructions.length-1) {
+            instructionCounter++;
+            sendCurrInstruction();
+        }
     } else if ('PREV_INSTRUCTION' in e.payload) {
-        instructionCounter++;
-        sendCurrInstruction();
+        if (instructionCounter > 0) {
+            instructionCounter--;
+            sendCurrInstruction();
+        }
     }
 });
